NS-115, give ImportASC the BackgroundJob treatment, with heavy lifting on external/asc
