+++ tests/test_api/test_job_controller.py
@@ -48,6 +48,7 @@ def post_basic_auth(client, path, credentials, data=None):
         '/api/job/create_sis_schema',
         '/api/job/generate_boac_analytics',
         '/api/job/generate_intermediate_tables',
+        '/api/job/import_asc_athletes',
         '/api/job/resync_canvas_snapshots',
         '/api/job/sync_canvas_snapshots',
         '/api/job/sync_file_to_s3',
@@ -78,6 +79,7 @@ class TestJobControllerAuthentication:
         '/api/job/create_sis_schema',
         '/api/job/generate_boac_analytics',
         '/api/job/generate_intermediate_tables',
+        '/api/job/import_asc_athletes',
         '/api/job/resync_canvas_snapshots',
         '/api/job/sync_canvas_snapshots',
     ],
